+++ src/libs/siyuan/protyle/util/functions.ts
@@ -1,0 +1,7 @@
+/**
+ * These functions are direct copies of not exposed functions from siyuan.
+ * @source app/src/protyle/util/functions.ts
+ */
+export const isMobile = () => {
+  return document.getElementById("sidebar") ? true : false;
+};
